GTBFS: use major version for packages
