+++ application/src/main/java/org/opentripplanner/updater/vehicle_rental/datasources/gbfs/v3/GbfsGeofencingZoneMapper.java
@@ -1,12 +1,15 @@
-package org.opentripplanner.updater.vehicle_rental.datasources.gbfs.v2_3;
+package org.opentripplanner.updater.vehicle_rental.datasources.gbfs.v3;
+
+import static org.opentripplanner.updater.vehicle_rental.datasources.gbfs.v3.GbfsFeedMapper.optionalLocalizedString;
 
 import java.util.List;
 import java.util.Objects;
+import javax.annotation.Nullable;
 import org.geojson.MultiPolygon;
-import org.mobilitydata.gbfs.v2_3.geofencing_zones.GBFSFeature;
-import org.mobilitydata.gbfs.v2_3.geofencing_zones.GBFSGeofencingZones;
+import org.mobilitydata.gbfs.v3_0.geofencing_zones.GBFSFeature;
+import org.mobilitydata.gbfs.v3_0.geofencing_zones.GBFSGeofencingZones;
+import org.mobilitydata.gbfs.v3_0.geofencing_zones.GBFSName;
 import org.opentripplanner.framework.i18n.I18NString;
-import org.opentripplanner.framework.i18n.NonLocalizedString;
 import org.opentripplanner.service.vehiclerental.model.GeofencingZone;
 
 /**
@@ -27,6 +30,7 @@ class GbfsGeofencingZoneMapper
       .getGeofencingZones()
       .getFeatures()
       .stream()
+      .filter(f -> f.getGeometry() != null)
       .map(this::toInternalModel)
       .filter(Objects::nonNull)
       .toList();
@@ -34,7 +38,7 @@ class GbfsGeofencingZoneMapper
 
   @Override
   protected boolean featureBansDropOff(GBFSFeature feature) {
-    return !feature.getProperties().getRules().get(0).getRideAllowed();
+    return !feature.getProperties().getRules().get(0).getRideEndAllowed();
   }
 
   @Override
@@ -48,7 +52,11 @@ class GbfsGeofencingZoneMapper
   }
 
   @Override
-  protected I18NString featureName(GBFSFeature feature) {
-    return NonLocalizedString.ofNullable(feature.getProperties().getName());
+  protected @Nullable I18NString featureName(GBFSFeature feature) {
+    return optionalLocalizedString(
+      feature.getProperties().getName(),
+      GBFSName::getLanguage,
+      GBFSName::getText
+    );
   }
 }
