+++ application/src/test/java/org/opentripplanner/updater/vehicle_rental/datasources/gbfs/v2/GbfsStationStatusMapperTest.java
@@ -1,5 +1,6 @@
-package org.opentripplanner.updater.vehicle_rental.datasources.gbfs.v3_0;
+package org.opentripplanner.updater.vehicle_rental.datasources.gbfs.v2;
 
+import static com.google.common.truth.Truth.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertSame;
@@ -10,7 +11,6 @@ import static org.opentripplanner.street.model.RentalFormFactor.BICYCLE;
 import static org.opentripplanner.street.model.RentalFormFactor.CAR;
 import static org.opentripplanner.street.model.RentalFormFactor.SCOOTER;
 import static org.opentripplanner.transit.model._data.TimetableRepositoryForTest.id;
-import static shadow.org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import java.util.Map;
@@ -19,9 +19,9 @@ import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
-import org.mobilitydata.gbfs.v3_0.station_status.GBFSStation;
-import org.mobilitydata.gbfs.v3_0.station_status.GBFSVehicleDocksAvailable;
-import org.mobilitydata.gbfs.v3_0.station_status.GBFSVehicleTypesAvailable;
+import org.mobilitydata.gbfs.v2_3.station_status.GBFSStation;
+import org.mobilitydata.gbfs.v2_3.station_status.GBFSVehicleDocksAvailable;
+import org.mobilitydata.gbfs.v2_3.station_status.GBFSVehicleTypesAvailable;
 import org.opentripplanner.service.vehiclerental.model.RentalVehicleType;
 import org.opentripplanner.service.vehiclerental.model.RentalVehicleTypeCount;
 import org.opentripplanner.service.vehiclerental.model.VehicleRentalStation;
@@ -42,7 +42,7 @@ class GbfsStationStatusMapperTest {
   void availableSpacesFromVehicles() {
     final var gbfsStation = station();
     gbfsStation.setNumDocksAvailable(3);
-    gbfsStation.setNumVehiclesAvailable(4);
+    gbfsStation.setNumBikesAvailable(4);
     var t = new GBFSVehicleTypesAvailable();
     t.setCount(1);
     t.setVehicleTypeId(TYPE_CAR.id().getId());
@@ -70,7 +70,7 @@ class GbfsStationStatusMapperTest {
   void availableSpacesFromTypes() {
     var gbfsStation = station();
     gbfsStation.setNumDocksAvailable(99);
-    gbfsStation.setNumVehiclesAvailable(4);
+    gbfsStation.setNumBikesAvailable(4);
     var type = new GBFSVehicleTypesAvailable();
     type.setCount(1);
     type.setVehicleTypeId(TYPE_CAR.id().getId());
@@ -106,7 +106,7 @@ class GbfsStationStatusMapperTest {
   void availableSpacesFromTypesWithoutAvailableVehicles() {
     var gbfsStation = station();
     gbfsStation.setNumDocksAvailable(88);
-    gbfsStation.setNumVehiclesAvailable(1);
+    gbfsStation.setNumBikesAvailable(1);
     var type = new GBFSVehicleTypesAvailable();
     type.setCount(0);
     type.setVehicleTypeId(TYPE_CAR.id().getId());
@@ -133,7 +133,7 @@ class GbfsStationStatusMapperTest {
   void noTypes() {
     var gbfsStation = station();
     gbfsStation.setNumDocksAvailable(1);
-    gbfsStation.setNumVehiclesAvailable(4);
+    gbfsStation.setNumBikesAvailable(4);
 
     var mapper = new GbfsStationStatusMapper(Map.of(ID, gbfsStation), Map.of());
 
@@ -152,7 +152,7 @@ class GbfsStationStatusMapperTest {
   void noSpaces(RentalFormFactor formFactor) {
     var gbfsStation = station();
     gbfsStation.setNumDocksAvailable(0);
-    gbfsStation.setNumVehiclesAvailable(4);
+    gbfsStation.setNumBikesAvailable(4);
 
     var mapper = new GbfsStationStatusMapper(Map.of(ID, gbfsStation), Map.of());
 
@@ -160,7 +160,7 @@ class GbfsStationStatusMapperTest {
 
     var spaces = mapped.vehicleSpaceCounts();
     assertEquals(0, spaces.total());
-    assertTrue(spaces.byType().isEmpty());
+    assertThat(spaces.byType()).isEmpty();
     assertEquals(Set.of(BICYCLE), mapped.formFactors());
     assertSame(ANY_TYPE, mapped.returnPolicy());
