+++ application/src/test/java/org/opentripplanner/updater/vehicle_rental/datasources/gbfs/v3/GbfsVehicleStatusMapperTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.updater.vehicle_rental.datasources.gbfs.v3_0;
+package org.opentripplanner.updater.vehicle_rental.datasources.gbfs.v3;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
